+++ benefits/WingspanBenefits/BenefitsService.cs
@@ -51,10 +51,10 @@ namespace WingspanBenefits
     {
         public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "4.2.0";
-        private const string _sdkGenVersion = "2.272.4";
+        private const string _sdkVersion = "4.2.1";
+        private const string _sdkGenVersion = "2.279.1";
         private const string _openapiDocVersion = "1.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 4.2.0 2.272.4 1.0.0 WingspanBenefits";
+        private const string _userAgent = "speakeasy-sdk/csharp 4.2.1 2.279.1 1.0.0 WingspanBenefits";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private Func<Security>? _securitySource;
@@ -66,7 +66,6 @@ namespace WingspanBenefits
             _serverUrl = serverUrl;
             SDKConfiguration = config;
         }
-        
 
         public async Task<GetBenefitsServiceResponse> GetBenefitsServiceAsync()
         {
@@ -106,7 +105,6 @@ namespace WingspanBenefits
             return response;
         }
 
-        
 
         public async Task<PatchBenefitsServiceIdResponse> PatchBenefitsServiceIdAsync(string id, ServiceEnablementUpdate? serviceEnablementUpdate = null)
         {
@@ -156,6 +154,5 @@ namespace WingspanBenefits
             return response;
         }
 
-        
     }
 }